Update README.md
Formik is a small library that helps you with the 3 most annoying parts:

*Getting values in and out of form state
*Validation and error messages
*Handling form submission
Yup is a schema builder for runtime value parsing and validation. Define a schema, transform a value to match, assert the shape of an existing value, or both. 
Yup schema are extremely expressive and allow modeling complex, interdependent validations, or value transformation.